+++ Pong/Assets/Scripts/UI/UIScore.cs
@@ -5,7 +5,7 @@ public class UIScore : MonoBehaviour
 {
     [SerializeField] private Text label;
 
-    public void UpdateScore(int score) {
+    public void UpdateScore(short score) {
         label.text = score.ToString();
     }
 }
